# user-server/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  hivemtk-user_internal_model.Customer:
    properties:
      churn_risk:
        type: string
      created_at:
        description: 时间
        type: string
      custom_attributes:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      douyin_open_id:
        type: string
      email:
        type: string
      feishu_open_id:
        type: string
      id:
        type: string
      kuaishou_open_id:
        type: string
      name:
        type: string
      owner_agent_id:
        type: integer
      phone:
        description: 强标识（用作 OneID 主键候选）
        type: string
      rfm_score:
        type: integer
      tags:
        description: 标签 + 评分
        type: string
      telegram_chat_id:
        type: integer
      telegram_username:
        type: string
      tiktok_open_id:
        type: string
      unified_id:
        type: string
      updated_at:
        type: string
      wechat_open_id:
        type: string
      wecom_external_id:
        type: string
      whatsapp_phone:
        type: string
      xianyu_id:
        type: string
      xiaohongshu_id:
        type: string
    type: object
  hivemtk-user_internal_model.CustomerEvent:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      event_data:
        type: string
      event_source:
        $ref: '#/definitions/hivemtk-user_internal_model.EventSource'
      event_type:
        $ref: '#/definitions/hivemtk-user_internal_model.EventType'
      id:
        type: string
      occurred_at:
        type: string
    type: object
...
  hivemtk-user_internal_model.EventSource:
    enum:
    - wechat
    - douyin
    - xiaohongshu
    - website
    - app
    type: string
    x-enum-varnames:
    - EventSourceWechat
    - EventSourceDouyin
    - EventSourceXiaohongshu
    - EventSourceWebsite
    - EventSourceApp
  hivemtk-user_internal_model.EventType:
    enum:
    - page_view
    - click
    - purchase
    - add_to_cart
    - signup
    - login
    type: string
    x-enum-varnames:
    - EventTypePageView
    - EventTypeClick
    - EventTypePurchase
    - EventTypeAddToCart
    - EventTypeSignup
    - EventTypeLogin
...
  hivemtk-user_internal_service.CustomerDTO:
    properties:
      douyin_open_id:
        type: string
      email:
        type: string
      phone:
        type: string
      wechat_open_id:
        type: string
      xiaohongshu_id:
        type: string
    type: object
  hivemtk-user_internal_service.CustomerProfile:
    properties:
      customer:
        $ref: '#/definitions/hivemtk-user_internal_model.Customer'
      recent_events:
        items:
          $ref: '#/definitions/hivemtk-user_internal_model.CustomerEvent'
        type: array
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: |-
    HiveMtk 用户端（私域独立部署）RESTful API。认证采用 JWT：请求头 Authorization: Bearer <token>。
    仅当 ENABLE_SWAGGER=true 时，通过 http://localhost:8204/swagger/index.html 本地访问。
  title: HiveMtk 用户端 API
  version: 3.41.0
paths:
...
  /api/customer:
    get:
      consumes:
      - application/json
      description: 获取所有客户，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取客户列表
      tags:
      - CDP-客户管理
    post:
      consumes:
      - application/json
      description: 创建新客户或更新现有客户（通过身份标识匹配）。至少需要提供一个有效身份标识。
      parameters:
      - description: 客户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hivemtk-user_internal_service.CustomerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            properties:
              data: {}
            type: object
        "400":
          description: 参数错误（缺身份标识 / 格式不合法）
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
      summary: 创建客户
      tags:
      - CDP-客户管理
...
  /api/customer/:id/tags:
    delete:
      consumes:
      - application/json
      description: 从客户移除一个或多个标签
      parameters:
      - description: 客户 ID
        in: path
        name: id
        required: true
        type: string
      - description: 标签列表
        in: body
        name: request
        required: true
        schema:
          properties:
            tags:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            properties:
              message:
                type: string
            type: object
      summary: 从客户移除标签
      tags:
      - CDP-客户管理
    post:
      consumes:
      - application/json
      description: 给客户添加一个或多个标签
      parameters:
      - description: 客户 ID
        in: path
        name: id
        required: true
        type: string
      - description: 标签列表
        in: body
        name: request
        required: true
        schema:
          properties:
            tags:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            properties:
              message:
                type: string
            type: object
      summary: 给客户添加标签
      tags:
      - CDP-客户管理
  /api/customer/{id}:
    get:
      consumes:
      - application/json
      description: 根据客户 ID 获取客户 360° 视图
      parameters:
      - description: 客户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            properties:
              data:
                $ref: '#/definitions/hivemtk-user_internal_service.CustomerProfile'
            type: object
      summary: 获取客户详情
      tags:
      - CDP-客户管理
  /api/customer/merge:
    post:
      consumes:
      - application/json
      description: 将 secondary 客户合并到 primary 客户，保留 primary 的信息并合并标签
      parameters:
      - description: 客户 ID
        in: body
        name: request
        required: true
        schema:
          properties:
            ' secondary_id':
              type: string
            primary_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 合并成功
          schema:
            properties:
              message:
                type: string
            type: object
      summary: 合并客户
      tags:
      - CDP-客户管理
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"